requirements.txt
absl-py==1.4.0
altair==4.2.2
astunparse==1.6.3
attrs==22.2.0
bleach==5.0.1
blinker==1.5
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
cmdtool==0.2
colorama==0.4.6
Cython==0.29.32
decorator==4.4.2
docopt==0.6.2
docutils==0.19
entrypoints==0.4
Flask==2.2.3
flatbuffers==23.3.3
fst-pso==1.8.1
FuzzyTM==2.0.5
gast==0.4.0
gensim==4.3.0
gevent==22.10.2
gitdb==4.0.10
GitPython==3.1.31
google-auth==2.16.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
greenlet==2.0.2
grpcio==1.51.3
h5py==3.8.0
idna==3.4
imageio==2.25.0
imageio-ffmpeg==0.4.8
importlib-metadata==6.0.0
itsdangerous==2.1.2
jaraco.classes==3.2.3
jax==0.4.6
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
keras==2.12.0
keyring==23.13.1
libclang==16.0.0
lxml==4.9.2
Markdown==3.4.3
markdown-it-py==2.1.0
MarkupSafe==2.1.2
mdurl==0.1.2
miniful==0.0.6
more-itertools==9.0.0
moviepy==1.0.3
numpy==1.23.5
oauthlib==3.2.2
opencv-python==4.7.0.68
opt-einsum==3.3.0
packaging==23.0
pandas==1.5.3
Pillow==9.4.0
pipreqs==0.4.11
pkginfo==1.9.6
proglog==0.1.10
protobuf==3.20.3
pyarrow==11.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydeck==0.8.0
pyFUME==0.2.25
pygame==2.1.2
Pygments==2.14.0
Pympler==1.0.1
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.1
pyrsistent==0.19.3
python-dateutil==2.8.2
pytube==12.1.2
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
pywin32-ctypes==0.2.0
readme-renderer==37.3
requests==2.28.2
requests-oauthlib==1.3.1
requests-toolbelt==0.10.1
rfc3986==2.0.0
rich==13.2.0
rsa==4.9
scipy==1.10.0
semver==2.13.0
simpful==2.9.0
six==1.16.0
smart-open==6.3.0
smmap==5.0.0
streamlit==1.20.0
tabulate==0.9.0
tensorboard==2.12.0
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-intel==2.12.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.2.0
toml==0.10.2
toolz==0.12.0
Topsis-Aashutosh-102053043==0.5
TOPSIS-GIRIK==1.1.0
Topsis-Pranav-102003432==0.6
Topsis-Sanjoli-102003425==0.5
tornado==6.2
tqdm==4.64.1
twine==4.0.2
typing_extensions==4.5.0
tzdata==2023.2
tzlocal==4.3
urllib3==1.26.14
utils==1.0.1
validators==0.20.0
watchdog==3.0.0
webencodings==0.5.1
Werkzeug==2.2.3
wget==3.2
wrapt==1.14.1
xlrd==2.0.1
yarg==0.1.9
zipp==3.11.0
zope.event==4.6
zope.interface==6.0
